improve MySQL storage + OpenWeatherMap temperature data
MySQL storage will now associate the location ID with the weather
data, too. Should have been this way from the start, because we
need to keep track which data belongs to which location.

Additionally, the OpenWeatherMap temperature data is now tweaked.
OpenWeatherMap usually returns temperature in Kelvin only, but
now the program automatically calculates the temperature in °C,
too.