update NULL values for temperature_F with proper calculated values
In version 0.8.0 the implementation for retrieving data with
OpenWeatherMap was changed so that the temperature in °F was
calculated before inserting data sets into the database. However,
this still left old data sets with NULL values for the
temperature, although those should have been updated at that time.
The new version of the updater program takes care of that.