implement parsing of snow data for DarkSky API
Previously, all precipitation data was considered to be rain.
However, DarkSky provides a JSON element that describes the type
of precipitation - rain or snow. This was not really noticed
before, because it is only present when precipitation is not
zero. Furthermore, all previous examples had it set to rain.

With that change we are almost halfway there to have snow
information in the weather data.