implement parsing of snow data for OpenWeatherMap API
Previously, all snow data was ignored and only rain data was
parsed. As a further improvement, empty rain and snow elements
are now interpreted as zero, because obviously the API seems to
say "I have some data here, but the value is zero, so I do not
write it explicitly." Otherwise the rain and snow elements are
completely absent, meaning there is no rain or snow data at all.

With that change we are almost there to have snow information in
the weather and forecast data.