fix for update mechanism to version 0.7.0
The update mechanism to version 0.7.0 will now also adjust the
wind direction for current weather entries of OpenWeatherMap that
have a non-integer value as wind direction. This change should
have been part of version 0.7.0, because the change that allowed
to parse not only integer values but also float values as wind
direction was introduced in version 0.7.0. (More precisely, it
was introduced in commit cbc43cea3f0a9bc7b8c463333af98e9244e907bd.
However, it was overlooked at that time.)