updater: only perform the update steps that are actually required
The update program looks at the database structure and uses it to
to determine the previous program version. This way the update is
much faster, because steps for older program versions are skipped.

If you want to have the old behaviour where all update steps are
applied, pass the parameter --full to the update program.