make checks for NaN in Weather's check for existing data explicit
Technically, checking for NaN via

    float f = ...; // assign some value
    bool isNotNaN = (f == f);

is valid and defined, because NaNs are the only floating-point
values that do not compare equal to themselves, even it the bit
representation is the same. That is guaranteed by the IEEE 754
standard which virtually every major C++ compiler uses for its
representation of floating-point types.

However, some people may be thrown off by this method and think
that the comparison is always true (which it is not for NaNs).
To avoid this, the NaN detection is now explicit.

If you do not known, what NaN is, or if you do know but want to
have some fun with JavaScript anyway, then type

    Array(16).join("wat" - 1) + " Batman!"

into the JavaScript interpreter of your choice. :)