tests: skip test for set environment variable in Windows
It seems that the SetEnvironmentVariable() method from the WinAPI
does not have any effect on variables that are queried by the
std::getenv() function. (Maybe changing the environment of the
current process needs some other approach on Windows, but I do
not have the time to investigate this further.)
Therefore, the test can never succeed, so it is justified to not
run that test section on Windows builds.