remove superfluous check for empty location in location parsing
Each location is checked for emptiness earlier on in the same
method, so empty locations are already caught at this point.
Checking again is unnecessary, because it will not yield any
different result, even if at that point the location information
may contain an additional country code, because country codes
alone are not enough to turn an "empty" location information
into a "non-empty" location information.