add Open-Meteo as fallback for task creation
Until now the task creator expected an API key for OpenWeatherMap
to get information about the location the user has entered.
However, if no such API key is set in the configuration file,
then it will use Open-Meteo's geocoding API instead. This works,
too, but currently still has the drawback that no current weather
data is available together with the location data. This may be
changed in a future commit.