fix problem with missing request time in Open-Meteo weather data
When current weather data was requested via the Open-Meteo API,
then the request time would be cleared (among with possible other
data) during the JSON parsing. This resulted in a missing request
time for the weather entry, which in turn caused any attempt to
save that data to the database to fail.

The commit fixes that problem, the request time is now preserved.