ci: update script for Debian package build to use latest commit
Previously, it used the latest tag, but that is outdated on
builds for untagged commit. So we use the current commit instead.
That's what we want in most cases, and for tagged commits the tag
and the commit hash point to the same Git tree anyway, so that
will not make any difference there.