adjust limit for Weatherstack's free plan to 100 requests per month
When we initially implemented requests to the Weatherstack API in
2019, the API allowed 1000 requests per month on the free plan.
This is the number that weather-information-collector has been
using ever since. However, Weatherstack changed the API limit to
250 requests per month in 2021 and again to 100 requests per
month in the second half of 2024.

So bascially, we can do between three and four API requests to
the Weatherstack API per day on the free plan. That way the free
plan is quite useless for most applications.